Refacted entire backend
