+++ src/server/api/beer-pong/enum.ts
@@ -1,5 +1,8 @@
 import { z } from "zod";
 
+/**
+ * The current status for a beer pong tournament
+ */
 export const BeerPongTournamentStatusSchema = z.enum([
   "CREATED",
   "ACTIVE",
@@ -13,6 +16,9 @@ export type BeerPongTournamentStatus = z.infer<
   typeof BeerPongTournamentStatusSchema
 >;
 
+/**
+ * The access level for a beer pong tournament, designated by the creator
+ */
 export const BeerPongTournamentAccessSchema = z.enum(["PUBLIC", "PIN"]);
 
 /**
