+++ src/server/api/beer-pong/team/controller/destroy.ts
@@ -1,0 +1,27 @@
+import { Controller } from "~/server/api/trpc";
+
+import { db } from "~/server/db";
+import { z } from "zod";
+import { protectedProcedure } from "../../../trpc";
+
+const InputSchema = z.object({
+  id: z.string(),
+});
+
+const OutputSchema = z.void();
+
+const handler: Controller<
+  z.infer<typeof InputSchema>,
+  z.infer<typeof OutputSchema>
+> = async ({ input, ctx }) => {
+  await db.beerPongTeam.delete({
+    where: {
+      id: input.id,
+    },
+  });
+};
+
+export default protectedProcedure
+  .input(InputSchema)
+  .output(OutputSchema)
+  .mutation(handler);
