+++ src/server/api/beer-pong/tournament/controller/start.ts
@@ -1,13 +1,27 @@
-import { db } from "~/server/db";
-import { TRPCError } from "@trpc/server";
-import { generateMatches } from "./generate_matches";
+import { Controller } from "~/server/api/trpc";
 
-export default async function startBeerPongTournament(
-  id: string,
-): Promise<void> {
+import { db } from "~/server/db";
+import { z } from "zod";
+import { protectedProcedure } from "~/server/api/trpc";
+import { TRPCError } from "@trpc/server";
+import { generateMatches } from "../helper/generate_matches";
+import { assertHasTournamentControl } from "../../middleware";
+
+const InputSchema = z.object({
+  tournamentId: z.string().cuid(),
+});
+
+const OutputSchema = z.void();
+
+const handler: Controller<
+  z.infer<typeof InputSchema>,
+  z.infer<typeof OutputSchema>
+> = async ({ input, ctx }) => {
+  assertHasTournamentControl({ ctx, tournamentId: input.tournamentId });
+
   const tournament = await db.beerPongTournament.findUniqueOrThrow({
     where: {
-      id,
+      id: input.tournamentId,
     },
     select: {
       status: true,
@@ -30,7 +44,7 @@ export default async function startBeerPongTournament(
 
   await db.beerPongTournament.update({
     where: {
-      id,
+      id: input.tournamentId,
     },
     data: {
       status: "ACTIVE",
@@ -40,11 +54,16 @@ export default async function startBeerPongTournament(
   await db.beerPongMatch.createMany({
     data: matches.map((match) => ({
       id: match.matchId,
-      tournamentId: id,
+      tournamentId: input.tournamentId,
       round: match.round,
       team1Id: match.team1Id,
       team2Id: match.team2Id,
       nextMatchId: match.nextMatchId,
     })),
   });
-}
+};
+
+export default protectedProcedure
+  .input(InputSchema)
+  .output(OutputSchema)
+  .mutation(handler);
