--- src/server/api/routers/questions.ts
@@ -1,7 +1,0 @@
-import { createTRPCRouter } from "~/server/api/trpc";
-
-export const questionsRouter = createTRPCRouter({
-  createGame: {},
-  getGame: {},
-  deleteGame: {},
-});
