--- src/server/service/beer-pong/team/create/schema.ts
@@ -1,9 +1,0 @@
-import { z } from "zod";
-
-export const CreateBeerPongTeamSchema = z.object({
-  tournamentId: z.string(),
-  teamName: z.string(),
-  joinTeam: z.boolean().default(true),
-});
-
-export type CreateBeerPongTeamInput = z.infer<typeof CreateBeerPongTeamSchema>;
