--- src/server/service/beer-pong/team/delete/mutation.ts
@@ -1,9 +1,0 @@
-import { db } from "~/server/db";
-
-export const deleteBeerPongTeamById = async (id: string) => {
-  await db.beerPongTeam.delete({
-    where: {
-      id,
-    },
-  });
-};
