--- src/server/service/beer-pong/team/get/query.ts
@@ -1,24 +1,0 @@
-import { db } from "~/server/db";
-
-export const getBeerPongTeamBYId = async (id: string) => {
-  const team = await db.beerPongTeam.findUniqueOrThrow({
-    where: {
-      id,
-    },
-    include: {
-      members: {
-        include: {
-          user: true,
-        },
-      },
-    },
-  });
-
-  return {
-    ...team,
-    members: team.members.map((mem) => ({
-      id: mem.userId,
-      nickname: mem.user.nickname,
-    })),
-  };
-};
