--- src/server/service/beer-pong/team/join/mutation.ts
@@ -1,21 +1,0 @@
-import { db } from "~/server/db";
-
-export const joinBeerPongTournament = async (
-  teamId: string,
-  userId: string,
-) => {
-  await db.userBeerPongTeam.create({
-    data: {
-      beerPongTeam: {
-        connect: {
-          id: teamId,
-        },
-      },
-      user: {
-        connect: {
-          id: userId,
-        },
-      },
-    },
-  });
-};
