--- src/server/service/beer-pong/tournament/delete/mutation.ts
@@ -1,11 +1,0 @@
-import { db } from "~/server/db";
-
-export const deleteBeerPongTournament = async (
-  tournamentId: string,
-): Promise<void> => {
-  await db.beerPongTournament.delete({
-    where: {
-      id: tournamentId,
-    },
-  });
-};
